fix: Move mixin in mixins.py and ajax view in ajax.py.
